Moved attendee, attendees from app
